add JWT claims

+++ resources/item.py
@@ -1,6 +1,6 @@
 from models.item import ItemModel
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 
 
 class Item(Resource):
@@ -41,6 +41,10 @@ class Item(Resource):
 
     @jwt_required()
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privileges required.'}
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
